requirements.txt
pandas
astunparse
typing
pathlib
python-magic
